+++ Pods/TunnelKit/TunnelKit/Sources/Protocols/OpenVPN/CryptoCBC.m
@@ -342,8 +342,8 @@ const NSInteger CryptoCBCMaxHMACLength = 100;
 {
     NSAssert(packet.length > 0, @"Decrypting an empty packet, how did it get this far?");
 
-    DATA_PATH_DECRYPT_INIT(packet.bytes)
-    if (packet.length < headerLength) {
+    DATA_PATH_DECRYPT_INIT(packet)
+    if (packet.length < headerLength + self.crypto.digestLength + self.crypto.cipherIVLength) {
         return NO;
     }